Add refresh session, exception handler in filter

+++ src/main/java/com/practice/shop/DAO/UserHasRoleRepository.java
@@ -3,7 +3,7 @@ package com.practice.shop.DAO;
 import com.practice.shop.models.UserActiveRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserHasRoleRepository extends JpaRepository<UserActiveRole, Long> {
+public interface UserHasRoleRepository extends JpaRepository<UserActiveRole, Integer> {
 
     boolean existsByUseridAndRoleId(Integer userid, Integer roleId);
 }
